# modules/rule-engine/fireflyframework-rule-engine-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-rule-engine
  description: Rule Engine Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/rules
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/rules/definitions/{id}:
    get:
      tags:
        - Rule Definitions
      summary: Get a rule definition by ID
      description: Retrieve a specific rule definition by its ID
      operationId: getRuleDefinitionById
      parameters:
        - name: id
          in: path
          description: Rule definition ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rule definition retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '404':
          description: Rule definition not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
    put:
      tags:
        - Rule Definitions
      summary: Update a rule definition
      description: Update an existing YAML DSL rule definition with validation
      operationId: updateRuleDefinition
      parameters:
        - name: id
          in: path
          description: Rule definition ID
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleDefinitionDTO'
        required: true
      responses:
        '200':
          description: Rule definition updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '400':
          description: Invalid rule definition or validation failed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '404':
          description: Rule definition not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
    delete:
      tags:
        - Rule Definitions
      summary: Delete a rule definition
      description: Delete a rule definition by ID
      operationId: deleteRuleDefinition
      parameters:
        - name: id
          in: path
          description: Rule definition ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Rule definition deleted successfully
        '404':
          description: Rule definition not found
        '500':
          description: Internal server error
...
  /api/v1/rules/definitions:
    post:
      tags:
        - Rule Definitions
      summary: Create a new rule definition
      description: 'Create and store a new YAML DSL rule definition with comprehensive validation. YAML content must follow naming conventions: camelCase for inputs (creditScore, annualIncome), snake_case for computed variables (debt_to_income, is_eligible), UPPER_CASE for constants (MIN_CREDIT_SCORE).'
      operationId: createRuleDefinition
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleDefinitionDTO'
        required: true
      responses:
        '201':
          description: Rule definition created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '400':
          description: Invalid rule definition, validation failed, or naming convention violations
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '409':
          description: Rule definition with the same code already exists
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
  /api/v1/rules/definitions/validate:
    post:
      tags:
        - Rule Definitions
      summary: Validate a YAML DSL rule definition
      description: Validate YAML DSL content without storing it. Checks syntax, naming conventions (camelCase inputs, snake_case computed, UPPER_CASE constants), dependencies, logic, and best practices.
      operationId: validateRuleDefinition
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: 'YAML DSL content to validate. Must follow naming conventions: camelCase for inputs, snake_case for computed variables, UPPER_CASE for constants.'
        required: true
      responses:
        '200':
          description: Validation completed (check response for validation results)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          description: Invalid YAML content or malformed request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ValidationResult'
  /api/v1/rules/definitions/filter:
    post:
      tags:
        - Rule Definitions
      summary: Get all rule definitions
      description: Retrieve rule definitions with optional filtering and pagination
      operationId: filterRuleDefinitions
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestRuleDefinitionDTO'
        required: true
      responses:
        '200':
          description: Rule definitions retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseRuleDefinitionDTO'
        '400':
          description: Invalid filter criteria
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseRuleDefinitionDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseRuleDefinitionDTO'
...
  /api/v1/rules/definitions/by-code/{code}:
    get:
      tags:
        - Rule Definitions
      summary: Get a rule definition by code
      description: Retrieve a specific rule definition by its unique code
      operationId: getRuleDefinitionByCode
      parameters:
        - name: code
          in: path
          description: Rule definition code
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rule definition retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '404':
          description: Rule definition not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RuleDefinitionDTO'
...
components:
  schemas:
    RuleDefinitionDTO:
      required:
        - code
        - isActive
        - name
        - yamlContent
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the rule definition
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        code:
          maxLength: 100
          minLength: 3
          pattern: ^[a-zA-Z][a-zA-Z0-9_]*$
          type: string
          description: Unique code identifier for the rule definition. Must start with a letter and contain only letters, numbers, and underscores.
          example: credit_scoring_v1
        name:
          maxLength: 200
          minLength: 3
          type: string
          description: Human-readable name for the rule definition
          example: Credit Scoring Rule v1
        description:
          maxLength: 1000
          minLength: 0
          type: string
          description: Detailed description of what this rule definition does
          example: Basic credit scoring rule for loan applications. Evaluates creditworthiness based on credit score, income, and debt ratios.
        yamlContent:
          maxLength: 50000
          minLength: 10
          type: string
          description: 'The YAML DSL content as a string. Must follow naming conventions: camelCase for inputs (creditScore, annualIncome), snake_case for computed variables (debt_to_income, is_eligible), UPPER_CASE for constants (MIN_CREDIT_SCORE).'
          example: |-
            name: "Credit Scoring"
            description: "Basic credit assessment"

            inputs:
              - creditScore        # camelCase input
              - annualIncome       # camelCase input
              - employmentYears    # camelCase input

            when:
              - creditScore at_least MIN_CREDIT_SCORE    # UPPER_CASE constant
              - annualIncome at_least 50000

            then:
              - calculate debt_to_income as existingDebt / annualIncome    # snake_case computed
              - set credit_tier to "PRIME"                                # snake_case computed
              - set is_eligible to true                                   # snake_case computed

            else:
              - set credit_tier to "STANDARD"
              - set is_eligible to false

            output:
              credit_tier: credit_tier
              is_eligible: is_eligible
              debt_to_income: debt_to_income
        version:
          pattern: ^\d+\.\d+\.\d+$
          type: string
          description: Version of the rule definition for tracking changes. Must follow semantic versioning format.
          example: 1.0.0
        isActive:
          type: boolean
          description: Whether this rule definition is currently active and can be used for evaluation
          example: true
        tags:
          maxLength: 500
          minLength: 0
          type: string
          description: Tags for categorizing and searching rule definitions (comma-separated)
          example: credit,scoring,loan,banking
        createdBy:
          type: string
          description: User who created this rule definition
          example: john.doe
        updatedBy:
          type: string
          description: User who last modified this rule definition
          example: jane.smith
        createdAt:
          type: string
          description: Timestamp when the rule definition was created
          format: date-time
          example: '2025-01-12T10:30:00Z'
        updatedAt:
          type: string
          description: Timestamp when the rule definition was last modified
          format: date-time
          example: '2025-01-12T15:45:00Z'
      description: Rule definition to create with YAML content following naming conventions
...
    ValidationIssue:
      type: object
      properties:
        code:
          type: string
        severity:
          type: string
          enum:
            - CRITICAL
            - ERROR
            - WARNING
            - INFO
        message:
          type: string
        description:
          type: string
        location:
          $ref: '#/components/schemas/ValidationLocation'
        suggestion:
          type: string
        examples:
          type: array
          items:
            type: string
    ValidationIssues:
      type: object
      properties:
        syntax:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        naming:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        logic:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        performance:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        bestPractices:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
    ValidationLocation:
      type: object
      properties:
        section:
          type: string
        lineNumber:
          type: integer
          format: int32
        path:
          type: string
        context:
          type: string
    ValidationMetadata:
      type: object
      properties:
        validatorVersion:
          type: string
        validatedAt:
          type: string
          format: date-time
        validationTimeMs:
          type: integer
          format: int64
        ruleName:
          type: string
        statistics:
          type: object
          additionalProperties:
            type: object
    ValidationResult:
      type: object
      properties:
        status:
          type: string
          enum:
            - VALID
            - WARNING
            - ERROR
            - CRITICAL_ERROR
        summary:
          $ref: '#/components/schemas/ValidationSummary'
        issues:
          $ref: '#/components/schemas/ValidationIssues'
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/ValidationSuggestion'
        metadata:
          $ref: '#/components/schemas/ValidationMetadata'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        valid:
          type: boolean
        errorSummary:
          type: string
    ValidationSuggestion:
      type: object
      properties:
        category:
          type: string
        title:
          type: string
        description:
          type: string
        example:
          type: string
        impact:
          type: integer
          format: int32
    ValidationSummary:
      type: object
      properties:
        totalIssues:
          type: integer
          format: int32
        criticalErrors:
          type: integer
          format: int32
        errors:
          type: integer
          format: int32
        warnings:
          type: integer
          format: int32
        suggestions:
          type: integer
          format: int32
        qualityScore:
          type: number
          format: double
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
    FilterRequestRuleDefinitionDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/RuleDefinitionDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria and pagination parameters
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponseRuleDefinitionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/RuleDefinitionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...